Use i128 for hylo-fix

+++ crates/cu_hylo_fix/tests/hylo_fix.rs
@@ -8,42 +8,42 @@ use cu_hylo_fix::hylo_fix::{
 };
 use svm_unit_test::svm_test;
 
-const MAX_BPS: i64 = 10_000;
+const MAX_BPS: i128 = 10_000;
 
-fn amount() -> Micro<i64> {
+fn amount() -> Micro<i128> {
     Micro::new(1_234_567_000_000)
 }
 
-fn scalar(value: i64) -> Micro<i64> {
+fn scalar(value: i128) -> Micro<i128> {
     Micro::new(value * 1_000_000)
 }
 
-fn bps(bps: i64) -> Micro<i64> {
+fn bps(bps: i128) -> Micro<i128> {
     Micro::new(bps * 1_000_000 / MAX_BPS)
 }
 
-fn checked_mul_at_n6(a: Micro<i64>, b: Micro<i64>) -> Option<Micro<i64>> {
+fn checked_mul_at_n6(a: Micro<i128>, b: Micro<i128>) -> Option<Micro<i128>> {
     a.checked_mul(&b).map(|product| product.convert::<N6>())
 }
 
-fn sqrt_newton(n: Micro<i64>, iterations: usize) -> Micro<i64> {
+fn sqrt_newton(n: Micro<i128>, iterations: usize) -> Micro<i128> {
     let mut guess = (n + scalar(1)) / 2;
 
     for _ in 0..iterations {
-        let quotient: Micro<i64> = (n / guess).convert::<N6>();
+        let quotient: Micro<i128> = (n / guess).convert::<N6>();
         guess = ((guess + quotient).convert::<N6>()) / 2;
     }
 
     guess
 }
 
-fn exp_taylor(x: Micro<i64>, terms: usize) -> Micro<i64> {
+fn exp_taylor(x: Micro<i128>, terms: usize) -> Micro<i128> {
     let mut sum = scalar(1);
     let mut term = scalar(1);
 
     let mut n = 1;
     while n <= terms {
-        term = (term * x).convert::<N6>() / n as i64;
+        term = (term * x).convert::<N6>() / n as i128;
         sum += term;
         n += 1;
     }
@@ -57,7 +57,7 @@ fn add_sub_mul_div() {
     let b = black_box(scalar(42));
     let c = black_box(scalar(7));
 
-    let out: Micro<i64> = (((a + b) - c) * bps(987)).convert::<N6>() / 3;
+    let out: Micro<i128> = (((a + b) - c) * bps(987)).convert::<N6>() / 3;
     black_box(out);
 }
 
@@ -77,9 +77,9 @@ fn checked_arithmetic() {
 #[svm_test]
 fn basis_points() {
     let principal = black_box(amount());
-    let fee: Micro<i64> = (principal * bps(30)).convert::<N6>();
-    let rebate: Micro<i64> = (principal * bps(5)).convert::<N6>();
-    let max_fee: Micro<i64> = (principal * bps(MAX_BPS)).convert::<N6>();
+    let fee: Micro<i128> = (principal * bps(30)).convert::<N6>();
+    let rebate: Micro<i128> = (principal * bps(5)).convert::<N6>();
+    let max_fee: Micro<i128> = (principal * bps(MAX_BPS)).convert::<N6>();
     let net = principal - fee + rebate;
 
     black_box((fee, rebate, max_fee, net));
@@ -88,8 +88,8 @@ fn basis_points() {
 #[svm_test]
 fn powers_and_sqrt() {
     let x = black_box(Micro::new(1_100_000));
-    let x2: Micro<i64> = (x * x).convert::<N6>();
-    let x4: Micro<i64> = (x2 * x2).convert::<N6>();
+    let x2: Micro<i128> = (x * x).convert::<N6>();
+    let x4: Micro<i128> = (x2 * x2).convert::<N6>();
     let root = sqrt_newton(black_box(scalar(144)), black_box(20));
 
     black_box((x2, x4, root));
